refactor openai logic to chat_utils
